Move minimal_training function to conftest of torch test subpackage
