+++ tests/influence/test_influences.py
@@ -4,13 +4,14 @@ from typing import Callable, Dict, NamedTuple, Tuple
 import numpy as np
 import pytest
 
+from .torch.conftest import minimal_training
+
 torch = pytest.importorskip("torch")
 
 import torch
 import torch.nn.functional as F
 from pytest_cases import fixture, parametrize, parametrize_with_cases
 from torch import nn
-from torch.optim import LBFGS
 from torch.utils.data import DataLoader, TensorDataset
 
 from pydvl.influence import InfluenceType, InversionMethod, compute_influences
@@ -18,50 +19,12 @@ from pydvl.influence.torch import TorchTwiceDifferentiable, model_hessian_low_ra
 
 from .conftest import (
     add_noise_to_linear_model,
-    linear_model, analytical_linear_influences,
+    analytical_linear_influences,
+    linear_model,
 )
 
 # Mark the entire module
 pytestmark = pytest.mark.torch
-
-
-def minimal_training(
-    model: torch.nn.Module,
-    dataloader: DataLoader,
-    loss_function: torch.nn.modules.loss._Loss,
-    lr: float = 0.01,
-    epochs: int = 50,
-):
-    """
-    Trains a PyTorch model using L-BFGS optimizer.
-
-    Args:
-        model: The PyTorch model to be trained.
-        dataloader: DataLoader providing the training data.
-        loss_function: The loss function to be used for training.
-        lr: The learning rate for the L-BFGS optimizer. Defaults to 0.01.
-        epochs: The number of training epochs. Defaults to 50.
-
-    Returns:
-        The trained model.
-    """
-    model = model.train()
-    optimizer = LBFGS(model.parameters(), lr=lr)
-
-    for epoch in range(epochs):
-        data = torch.cat([inputs for inputs, targets in dataloader])
-        targets = torch.cat([targets for inputs, targets in dataloader])
-
-        def closure():
-            optimizer.zero_grad()
-            outputs = model(data)
-            loss = loss_function(outputs, targets)
-            loss.backward()
-            return loss
-
-        optimizer.step(closure)
-
-    return model
 
 
 def create_conv3d_nn():
